refactor: Move `scrapers` module from `praice/data_handling` to `praice/libs`
